pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fuk"
version = "0.1.0"
description = "Framework for Unified Kreation - Local-first rendering pipeline for professional image and video generation"
authors = [
    {name = "Brad", email = "hello@castlegreyscale.com"}
]
requires-python = ">=3.10,<3.11"
license = {text = "TBD"}
readme = "README.md"
keywords = ["ai", "image-generation", "local-first", "rendering", "pipeline"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
]

# Minimal dependencies - musubi handles ML/AI stack
dependencies = [
	 "accelerate==1.6.0",
    "av==14.0.1",
    "bitsandbytes",
    "diffusers==0.32.1",
    "einops==0.7.0",
    "huggingface-hub==0.34.3",
    "opencv-python==4.10.0.84",
    "pillow>=11.3.0",
    "safetensors==0.4.5",
    # "sageattention>=1.0.6",
    "toml==0.10.2",
    "tqdm==4.67.1",
    "transformers==4.54.1",
    "voluptuous==0.15.2",
    # Wan2.1
    "ftfy==6.3.1",
    "easydict==1.13",
    # FLUX.1 Kontext
    "sentencepiece==0.2.1",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "gradio>=4.0.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "ffmpeg>=1.4",
    "openEXR>=3.4.4",
    "Imath>=0.0.2",
    "python-multipart>=0.0.6",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
cu124 = [
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]
cu128 = [
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]

[project.urls]
Homepage = "https://github.com/castlegreyscale/fuk"
Repository = "https://github.com/castlegreyscale/fuk"

[tool.hatchling.build.targets.wheel]
packages = ["fuk"]

# REQUIRED: Allow git dependencies
[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I"]
ignore = ["E501"]
